Add hashing of input data
This commit introduces hashing of the image input data to the step settings.

When loading the step settings, we now generate a hash of the image input data and check it against the hash of the input data that was used last time this step was opened. The remaining settings and previous output file will only be reloaded if the current input data matches the previous input data.

The hash of the image input data is saved to the step settings when closing the step.